+++ cypress/integration/authorization.spec.ts
@@ -1,3 +1,47 @@
+function createACLSparql(document: string): string {
+    return `
+        INSERT DATA {
+            @prefix acl: <http://www.w3.org/ns/auth/acl#> .
+
+            <#owner>
+                a acl:Authorization ;
+                acl:agent <http://localhost:4000/alice/profile/card#me>, <mailto:alice@example.com> ;
+                acl:accessTo <http://localhost:4000/alice/cookbook/${document}> ;
+                acl:mode acl:Read, acl:Write, acl:Control .
+        }
+    `;
+}
+
+function publishACLSparql(document: string): string {
+    return `
+        INSERT DATA {
+            @prefix acl: <http://www.w3.org/ns/auth/acl#> .
+            @prefix foaf: <http://xmlns.com/foaf/0.1/> .
+
+            <#public>
+                a acl:Authorization ;
+                acl:agentClass foaf:Agent ;
+                acl:accessTo <http://localhost:4000/alice/cookbook/${document}> ;
+                acl:mode acl:Read .
+        }
+    `;
+}
+
+function unpublishACLSparql(document: string): string {
+    return `
+        DELETE DATA {
+            @prefix acl: <http://www.w3.org/ns/auth/acl#> .
+            @prefix foaf: <http://xmlns.com/foaf/0.1/> .
+
+            <#public>
+                a acl:Authorization ;
+                acl:agentClass foaf:Agent ;
+                acl:accessTo <http://localhost:4000/alice/cookbook/${document}> ;
+                acl:mode acl:Read .
+        }
+    `;
+}
+
 describe('Authorization', () => {
 
     beforeEach(() => {
@@ -33,91 +77,58 @@ describe('Authorization', () => {
 
         // Assert
         cy.see('updating permissions');
-        cy.see('Public', 'button');
+        cy.see('Public', 'button', { timeout: 10000 });
         cy.dontSee('This recipe is private');
 
         cy.get('@createTypeIndex').should('not.be.null');
-        cy.get('@createPublicList').its('request.body').should('be.sparql', `
-            INSERT DATA {
-                @prefix schema: <https://schema.org/> .
-                @prefix purl: <http://purl.org/dc/terms/> .
-
-                <#it>
-                    a schema:ItemList ;
-                    schema:name "Public Recipes" ;
-                    purl:creator <http://localhost:4000/alice/profile/card#me> ;
-                    schema:itemListElement <#[[item][%uuid%]]> .
-
-                <#[[item][%uuid%]]>
-                    a schema:ListItem ;
-                    schema:item <http://localhost:4000/alice/cookbook/ramen#it> .
-            }
-        `);
-        cy.get('@updateRamen').its('request.body').should('be.sparql', `
-            DELETE DATA {
-                @prefix crdt: <https://vocab.noeldemartin.com/crdt/> .
-                @prefix xsd: <http://www.w3.org/2001/XMLSchema#> .
-
-                <#it-metadata> crdt:updatedAt "[[createdAt][.*]]"^^xsd:dateTime .
-            } ;
-            INSERT DATA {
-                @prefix schema: <https://schema.org/> .
-                @prefix purl: <http://purl.org/dc/terms/> .
-                @prefix crdt: <https://vocab.noeldemartin.com/crdt/> .
-                @prefix xsd: <http://www.w3.org/2001/XMLSchema#> .
-
-                <#it> purl:isReferencedBy <http://localhost:4000/alice/cookbook/public#it> .
-                <#it-metadata> crdt:updatedAt "[[updatedAt][.*]]"^^xsd:dateTime .
-
-                <#it-operation-[[name][%uuid%]]>
-                    a crdt:SetPropertyOperation ;
-                    crdt:resource <#it> ;
-                    crdt:date "[[createdAt][.*]]"^^xsd:dateTime ;
-                    crdt:property schema:name ;
-                    crdt:value "Ramen" .
-
-                <#it-operation-[[image][%uuid%]]>
-                    a crdt:SetPropertyOperation ;
-                    crdt:resource <#it> ;
-                    crdt:date "[[createdAt][.*]]"^^xsd:dateTime ;
-                    crdt:property schema:image ;
-                    crdt:value <http://localhost:4000/alice/cookbook/ramen.png> .
-
-                <#it-operation-[[listing][%uuid%]]>
-                    a crdt:AddPropertyOperation ;
-                    crdt:resource <#it> ;
-                    crdt:date "[[updatedAt][.*]]"^^xsd:dateTime ;
-                    crdt:property purl:isReferencedBy ;
-                    crdt:value <http://localhost:4000/alice/cookbook/public#it> .
-            }
-        `);
+        cy.fixture('create-recipes-list.sparql').then(sparql => {
+            cy.get('@createPublicList').its('request.body').should('be.sparql', sparql);
+        });
+        cy.fixture('publish-ramen.sparql').then(sparql => {
+            cy.get('@updateRamen').its('request.body').should('be.sparql', sparql);
+        });
 
         publicDocuments.forEach(document => {
             cy.get(`@patch-${document}-ACL.all`).should('have.length', 2);
-            cy.get(`@patch-${document}-ACL.0`).its('request.body').should('be.sparql', `
-                INSERT DATA {
-                    @prefix acl: <http://www.w3.org/ns/auth/acl#> .
-
-                    <#owner>
-                        a acl:Authorization ;
-                        acl:agent <http://localhost:4000/alice/profile/card#me>, <mailto:alice@example.com> ;
-                        acl:accessTo <http://localhost:4000/alice/cookbook/${document}> ;
-                        acl:mode acl:Read, acl:Write, acl:Control .
-                }
-            `);
-            cy.get(`@patch-${document}-ACL.1`).its('request.body').should('be.sparql', `
-                INSERT DATA {
-                    @prefix acl: <http://www.w3.org/ns/auth/acl#> .
-                    @prefix foaf: <http://xmlns.com/foaf/0.1/> .
-
-                    <#public>
-                        a acl:Authorization ;
-                        acl:agentClass foaf:Agent ;
-                        acl:accessTo <http://localhost:4000/alice/cookbook/${document}> ;
-                        acl:mode acl:Read .
-                }
-            `);
+            cy.get(`@patch-${document}-ACL.0`).its('request.body').should('be.sparql', createACLSparql(document));
+            cy.get(`@patch-${document}-ACL.1`).its('request.body').should('be.sparql', publishACLSparql(document));
         });
+    });
+
+    it('Publishes unlisted recipes', () => {
+        // Arrange
+        const publicDocuments = ['ramen', 'ramen.png'];
+
+        cy.intercept('PATCH', 'http://localhost:4000/alice/settings/publicTypeIndex').as('createTypeIndex');
+        cy.intercept('PATCH', 'http://localhost:4000/alice/cookbook/public').as('createPublicList');
+        cy.intercept('PATCH', 'http://localhost:4000/alice/cookbook/ramen').as('updateRamen');
+
+        publicDocuments.forEach(acl => {
+            cy.intercept('PATCH', `http://localhost:4000/alice/cookbook/${acl}.acl`).as(`patch-${acl}-ACL`);
+        });
+
+        // Act
+        cy.press('Ramen');
+        cy.press('Share');
+        cy.see('Private', 'button');
+        cy.see('This recipe is private');
+        cy.press('Private', 'button');
+        cy.press('Unlisted', 'li');
+
+        // Assert
+        cy.see('updating permissions');
+        cy.see('Unlisted', 'button', { timeout: 10000 });
+        cy.dontSee('This recipe is private');
+
+        publicDocuments.forEach(document => {
+            cy.get(`@patch-${document}-ACL.all`).should('have.length', 2);
+            cy.get(`@patch-${document}-ACL.0`).its('request.body').should('be.sparql', createACLSparql(document));
+            cy.get(`@patch-${document}-ACL.1`).its('request.body').should('be.sparql', publishACLSparql(document));
+        });
+
+        cy.get('@createTypeIndex').should('be.null');
+        cy.get('@createPublicList').should('be.null');
+        cy.get('@updateRamen').should('be.null');
     });
 
     it('Makes public recipes private', () => {
@@ -150,59 +161,14 @@ describe('Authorization', () => {
 
         // Assert
         cy.see('updating permissions');
-        cy.see('Private');
+        cy.see('Private', { timeout: 10000 });
         cy.see('This recipe is private');
 
-        cy.get('@patchRamenACL').its('request.body').should('be.sparql', `
-            DELETE DATA {
-                @prefix acl: <http://www.w3.org/ns/auth/acl#> .
-                @prefix foaf: <http://xmlns.com/foaf/0.1/> .
-
-                <#public>
-                    a acl:Authorization ;
-                    acl:agentClass foaf:Agent ;
-                    acl:accessTo <http://localhost:4000/alice/cookbook/ramen> ;
-                    acl:mode acl:Read .
-            }
-        `);
-        cy.get('@patchImageACL').its('request.body').should('be.sparql', `
-            DELETE DATA {
-                @prefix acl: <http://www.w3.org/ns/auth/acl#> .
-                @prefix foaf: <http://xmlns.com/foaf/0.1/> .
-
-                <#public>
-                    a acl:Authorization ;
-                    acl:agentClass foaf:Agent ;
-                    acl:accessTo <http://localhost:4000/alice/cookbook/ramen.png> ;
-                    acl:mode acl:Read .
-            }
-        `);
-        cy.get('@patchRamen').its('request.body').should('be.sparql', `
-            DELETE DATA {
-                @prefix schema: <https://schema.org/> .
-                @prefix purl: <http://purl.org/dc/terms/> .
-                @prefix crdt: <https://vocab.noeldemartin.com/crdt/> .
-                @prefix xsd: <http://www.w3.org/2001/XMLSchema#> .
-
-                <#it> purl:isReferencedBy <http://localhost:4000/alice/cookbook/public#it> .
-                <#it-metadata> crdt:updatedAt "[[.*]]"^^xsd:dateTime .
-            } ;
-            INSERT DATA {
-                @prefix schema: <https://schema.org/> .
-                @prefix purl: <http://purl.org/dc/terms/> .
-                @prefix crdt: <https://vocab.noeldemartin.com/crdt/> .
-                @prefix xsd: <http://www.w3.org/2001/XMLSchema#> .
-
-                <#it-metadata> crdt:updatedAt "[[updatedAt][.*]]"^^xsd:dateTime .
-
-                <#it-operation-[[listing][%uuid%]]>
-                    a crdt:RemovePropertyOperation ;
-                    crdt:resource <#it> ;
-                    crdt:date "[[updatedAt][.*]]"^^xsd:dateTime ;
-                    crdt:property purl:isReferencedBy ;
-                    crdt:value <http://localhost:4000/alice/cookbook/public#it> .
-            }
-        `);
+        cy.get('@patchRamenACL').its('request.body').should('be.sparql', unpublishACLSparql('ramen'));
+        cy.get('@patchImageACL').its('request.body').should('be.sparql', unpublishACLSparql('ramen.png'));
+        cy.fixture('unpublish-ramen.sparql').then(sparql => {
+            cy.get('@patchRamen').its('request.body').should('be.sparql', sparql);
+        });
         cy.get('@patchList').its('request.body').should('be.sparql', `
             DELETE DATA {
                 @prefix schema: <https://schema.org/> .
@@ -241,71 +207,14 @@ describe('Authorization', () => {
 
         // Assert
         cy.see('updating permissions');
-        cy.see('Public', 'button');
+        cy.see('Public', 'button', { timeout: 10000 });
         cy.dontSee('This recipe is private');
 
-        cy.get('@patchRamenACL').its('request.body').should('be.sparql', `
-            INSERT DATA {
-                @prefix acl: <http://www.w3.org/ns/auth/acl#> .
-                @prefix foaf: <http://xmlns.com/foaf/0.1/> .
-
-                <#public>
-                    a acl:Authorization ;
-                    acl:agentClass foaf:Agent ;
-                    acl:accessTo <http://localhost:4000/alice/cookbook/ramen> ;
-                    acl:mode acl:Read .
-            }
-        `);
-        cy.get('@patchImageACL').its('request.body').should('be.sparql', `
-            INSERT DATA {
-                @prefix acl: <http://www.w3.org/ns/auth/acl#> .
-                @prefix foaf: <http://xmlns.com/foaf/0.1/> .
-
-                <#public>
-                    a acl:Authorization ;
-                    acl:agentClass foaf:Agent ;
-                    acl:accessTo <http://localhost:4000/alice/cookbook/ramen.png> ;
-                    acl:mode acl:Read .
-            }
-        `);
-        cy.get('@patchRamen').its('request.body').should('be.sparql', `
-            DELETE DATA {
-                @prefix crdt: <https://vocab.noeldemartin.com/crdt/> .
-                @prefix xsd: <http://www.w3.org/2001/XMLSchema#> .
-
-                <#it-metadata> crdt:updatedAt "[[createdAt][.*]]"^^xsd:dateTime .
-            } ;
-            INSERT DATA {
-                @prefix schema: <https://schema.org/> .
-                @prefix purl: <http://purl.org/dc/terms/> .
-                @prefix crdt: <https://vocab.noeldemartin.com/crdt/> .
-                @prefix xsd: <http://www.w3.org/2001/XMLSchema#> .
-
-                <#it> purl:isReferencedBy <http://localhost:4000/alice/cookbook/public#it> .
-                <#it-metadata> crdt:updatedAt "[[updatedAt][.*]]"^^xsd:dateTime .
-
-                <#it-operation-[[name][%uuid%]]>
-                    a crdt:SetPropertyOperation ;
-                    crdt:resource <#it> ;
-                    crdt:date "[[createdAt][.*]]"^^xsd:dateTime ;
-                    crdt:property schema:name ;
-                    crdt:value "Ramen" .
-
-                <#it-operation-[[image][%uuid%]]>
-                    a crdt:SetPropertyOperation ;
-                    crdt:resource <#it> ;
-                    crdt:date "[[createdAt][.*]]"^^xsd:dateTime ;
-                    crdt:property schema:image ;
-                    crdt:value <http://localhost:4000/alice/cookbook/ramen.png> .
-
-                <#it-operation-[[listing][%uuid%]]>
-                    a crdt:AddPropertyOperation ;
-                    crdt:resource <#it> ;
-                    crdt:date "[[updatedAt][.*]]"^^xsd:dateTime ;
-                    crdt:property purl:isReferencedBy ;
-                    crdt:value <http://localhost:4000/alice/cookbook/public#it> .
-            }
-        `);
+        cy.get('@patchRamenACL').its('request.body').should('be.sparql', publishACLSparql('ramen'));
+        cy.get('@patchImageACL').its('request.body').should('be.sparql', publishACLSparql('ramen.png'));
+        cy.fixture('publish-ramen.sparql').then(sparql => {
+            cy.get('@patchRamen').its('request.body').should('be.sparql', sparql);
+        });
         cy.get('@patchList').its('request.body').should('be.sparql', `
             INSERT DATA {
                 @prefix schema: <https://schema.org/> .